10787: Extract hook from Full Address Search (#9163)
* changelog: Internal, In-Person Proofing, Extract FAS hook

* Add simple test for hook
