+++ app/javascript/packages/address-search/index.tsx
@@ -3,6 +3,7 @@ import InPersonLocations from './components/in-person-locations';
 import AddressInput from './components/address-input';
 import AddressSearch from './components/address-search';
 import NoInPersonLocationsDisplay from './components/no-in-person-locations-display';
+import { requestUspsLocations } from './hooks/use-usps-locations';
 
 export {
   snakeCase,
@@ -11,6 +12,7 @@ export {
   InPersonLocations,
   AddressInput,
   NoInPersonLocationsDisplay,
+  requestUspsLocations,
 };
 
 export default AddressSearch;
