+++ app/javascript/packages/document-capture/components/in-person-full-address-search.tsx
@@ -1,139 +1,13 @@
 import { TextInput, SelectInput } from '@18f/identity-components';
 import { useState, useRef, useEffect, useCallback, useContext } from 'react';
 import { t } from '@18f/identity-i18n';
-import { request } from '@18f/identity-request';
 import ValidatedField from '@18f/identity-validated-field/validated-field';
 import SpinnerButton, { SpinnerButtonRefHandle } from '@18f/identity-spinner-button/spinner-button';
 import type { RegisterFieldCallback } from '@18f/identity-form-steps';
-import useSWR from 'swr/immutable';
 import { useDidUpdateEffect } from '@18f/identity-react-hooks';
-import { transformKeys, snakeCase } from '@18f/identity-address-search';
-import type {
-  LocationQuery,
-  PostOffice,
-  FormattedLocation,
-} from '@18f/identity-address-search/types';
+import type { LocationQuery, FormattedLocation } from '@18f/identity-address-search/types';
 import { InPersonContext } from '../context';
-
-const formatLocations = (postOffices: PostOffice[]): FormattedLocation[] =>
-  postOffices.map((po: PostOffice, index) => ({
-    formattedCityStateZip: `${po.city}, ${po.state}, ${po.zip_code_5}-${po.zip_code_4}`,
-    id: index,
-    distance: po.distance,
-    name: po.name,
-    saturdayHours: po.saturday_hours,
-    streetAddress: po.address,
-    sundayHours: po.sunday_hours,
-    weekdayHours: po.weekday_hours,
-    isPilot: !!po.is_pilot,
-  }));
-
-const requestUspsLocations = async ({
-  address,
-  locationsURL,
-}: {
-  locationsURL: string;
-  address: LocationQuery;
-}): Promise<FormattedLocation[]> => {
-  const response = await request<PostOffice[]>(locationsURL, {
-    method: 'post',
-    json: { address: transformKeys(address, snakeCase) },
-  });
-
-  return formatLocations(response);
-};
-
-function useUspsLocations(locationsURL: string) {
-  const [locationQuery, setLocationQuery] = useState<LocationQuery | null>(null);
-  const validatedAddressFieldRef = useRef<HTMLFormElement>(null);
-  const validatedCityFieldRef = useRef<HTMLFormElement>(null);
-  const validatedStateFieldRef = useRef<HTMLFormElement>(null);
-  const validatedZipCodeFieldRef = useRef<HTMLFormElement>(null);
-
-  const checkValidityAndDisplayErrors = (address, city, state, zipCode) => {
-    let formIsValid = true;
-    const zipCodeIsValid = zipCode.length === 5 && !!zipCode.match(/\d{5}/);
-
-    if (address.length === 0) {
-      validatedAddressFieldRef.current?.setCustomValidity(t('simple_form.required.text'));
-      formIsValid = false;
-    } else {
-      validatedAddressFieldRef.current?.setCustomValidity('');
-    }
-
-    if (city.length === 0) {
-      formIsValid = false;
-      validatedCityFieldRef.current?.setCustomValidity(t('simple_form.required.text'));
-    } else {
-      validatedCityFieldRef.current?.setCustomValidity('');
-    }
-
-    if (state.length === 0) {
-      formIsValid = false;
-      validatedStateFieldRef.current?.setCustomValidity(t('simple_form.required.text'));
-    } else {
-      validatedStateFieldRef.current?.setCustomValidity('');
-    }
-
-    if (zipCode.length === 0) {
-      formIsValid = false;
-      validatedZipCodeFieldRef.current?.setCustomValidity(t('simple_form.required.text'));
-    } else {
-      validatedZipCodeFieldRef.current?.setCustomValidity('');
-    }
-
-    validatedAddressFieldRef.current?.reportValidity();
-    validatedCityFieldRef.current?.reportValidity();
-    validatedStateFieldRef.current?.reportValidity();
-    validatedZipCodeFieldRef.current?.reportValidity();
-
-    return formIsValid && zipCodeIsValid;
-  };
-
-  const handleLocationSearch = useCallback(
-    (event, addressValue, cityValue, stateValue, zipCodeValue) => {
-      event.preventDefault();
-      const address = addressValue.trim();
-      const city = cityValue.trim();
-      const zipCode = zipCodeValue.trim();
-
-      const formIsValid = checkValidityAndDisplayErrors(address, city, stateValue, zipCode);
-
-      if (!formIsValid) {
-        return;
-      }
-
-      setLocationQuery({
-        address: `${address}, ${city}, ${stateValue} ${zipCode}`,
-        streetAddress: address,
-        city,
-        state: stateValue,
-        zipCode,
-      });
-    },
-    [],
-  );
-
-  const {
-    data: locationResults,
-    isLoading: isLoadingLocations,
-    error: uspsError,
-  } = useSWR([locationQuery], ([address]) =>
-    address ? requestUspsLocations({ address, locationsURL }) : null,
-  );
-
-  return {
-    locationQuery,
-    locationResults,
-    uspsError,
-    isLoading: isLoadingLocations,
-    handleLocationSearch,
-    validatedAddressFieldRef,
-    validatedCityFieldRef,
-    validatedStateFieldRef,
-    validatedZipCodeFieldRef,
-  };
-}
+import useValidatedUspsLocations from '../hooks/use-validated-usps-locations';
 
 interface FullAddressSearchProps {
   registerField?: RegisterFieldCallback;
@@ -170,7 +44,7 @@ function FullAddressSearch({
     validatedCityFieldRef,
     validatedStateFieldRef,
     validatedZipCodeFieldRef,
-  } = useUspsLocations(locationsURL);
+  } = useValidatedUspsLocations(locationsURL);
 
   const inputChangeHandler =
     <T extends HTMLElement & { value: string }>(input) =>
